move _context.py to private
